AStyle: don't indent cases
This was a bug in our configuration. Earlier versions of our style did not
specify this option. We found out for some code in our codebase that it was
changed because of the setting.

An example:

```
case: {
    // code here
}

//would be changed to

case: {
    // code here
   }
```

The first one is correct and do not need to be indented (it's enough switch is
indented).